Made an error exception

+++ src/main/java/org/example/JobSearch/service/impl/VacancyServiceImpl.java
@@ -2,6 +2,7 @@ package org.example.JobSearch.service.impl;
 
 import org.example.JobSearch.dao.VacancyDao;
 import org.example.JobSearch.dto.VacancyDTO;
+import org.example.JobSearch.exceptions.VacancyNotFoundException;
 import org.example.JobSearch.model.Vacancy;
 import org.example.JobSearch.service.VacancyService;
 import lombok.RequiredArgsConstructor;
@@ -17,7 +18,11 @@ public class VacancyServiceImpl implements VacancyService {
 
     @Override
     public List<VacancyDTO> getVacanciesByCategory(Long categoryId) {
-        return vacancyDao.getVacanciesByCategory(categoryId).stream().map(this::toDTO).collect(Collectors.toList());
+        List<VacancyDTO> vacancies = vacancyDao.getVacanciesByCategory(categoryId).stream().map(this::toDTO).collect(Collectors.toList());
+        if (vacancies.isEmpty()) {
+            throw new VacancyNotFoundException("No vacancies found for category ID: " + categoryId);
+        }
+        return vacancies;
     }
 
     @Override
@@ -27,25 +32,39 @@ public class VacancyServiceImpl implements VacancyService {
 
     @Override
     public void updateVacancy(Long vacancyId, VacancyDTO vacancyDto) {
+        if (vacancyDao.getVacancyById(vacancyId) == null) {
+            throw new VacancyNotFoundException("Vacancy not found with ID: " + vacancyId);
+        }
         vacancyDao.updateVacancy(vacancyId, vacancyDto);
     }
 
     @Override
     public void deleteVacancy(Long vacancyId) {
+        if (vacancyDao.getVacancyById(vacancyId) == null) {
+            throw new VacancyNotFoundException("Vacancy not found with ID: " + vacancyId);
+        }
         vacancyDao.deleteVacancy(vacancyId);
     }
 
     @Override
     public List<VacancyDTO> getAllVacancies() {
-        return vacancyDao.getAllVacancies().stream().map(this::toDTO).collect(Collectors.toList());
+        List<VacancyDTO> vacancies = vacancyDao.getAllVacancies().stream().map(this::toDTO).collect(Collectors.toList());
+        if (vacancies.isEmpty()) {
+            throw new VacancyNotFoundException("No vacancies found");
+        }
+        return vacancies;
     }
 
     @Override
     public List<VacancyDTO> getRespApplToVacancy(Long applicantId) {
-        return vacancyDao.getRespondedVacancies(applicantId)
+        List<VacancyDTO> vacancies = vacancyDao.getRespondedVacancies(applicantId)
                 .stream()
                 .map(this::toDTO)
                 .collect(Collectors.toList());
+        if (vacancies.isEmpty()) {
+            throw new VacancyNotFoundException("No responded vacancies found for applicant ID: " + applicantId);
+        }
+        return vacancies;
     }
 
     private VacancyDTO toDTO(Vacancy vacancy) {
